// src/app/models/hourly.ts
export interface Hourly {
  apparentTemperature: number;
  cardinal: string;
  cloudCover: number;
  dewPoint: number;
  humidity: number;
  icon: string;
  ozone: number;
  precipIntensity: number;
  precipProbability: number;
  pressure: number;
  summary: string;
  tempC: number;
  temperature: number;
  time: number;
  uvIndex: number;
  visibility: number;
  windBearing: number;
  windGust: number;
  windKph: number;
  windSpeed: number;
}